@@ -1,25 +1,23 @@
 package com.bubuwork.jf.om.entity;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+
+import jakarta.persistence.*;
+import lombok.Data;
 
 @Entity(name="GENERIC_ORDER_ITEM")
 @Table
+@Data
 public class GenericOrderItem {
   
   @Id
-  @GeneratedValue(strategy=GenerationType.AUTO)
+  @GeneratedValue(strategy= GenerationType.AUTO)
   @Column(name="ID")
   private Long id;
   
   @Column(name="ORDER_ID")
   private Long orderId;
   
-  @Column(name="DESCRIPTIOIN")
+  @Column(name="DESCRIPTION")
   private String description;
   
   @Column(name="UNIT")
@@ -28,73 +26,5 @@ public class GenericOrderItem {
   @Column(name="PRICE")
   private double price;
 
-  /**
-   * @return the id
-   */
-  public Long getId() {
-    return id;
-  }
 
-  /**
-   * @param id the id to set
-   */
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  /**
-   * @return the orderId
-   */
-  public Long getOrderId() {
-    return orderId;
-  }
-
-  /**
-   * @param orderId the orderId to set
-   */
-  public void setOrderId(Long orderId) {
-    this.orderId = orderId;
-  }
-
-  /**
-   * @return the description
-   */
-  public String getDescription() {
-    return description;
-  }
-
-  /**
-   * @param description the description to set
-   */
-  public void setDescription(String description) {
-    this.description = description;
-  }
-
-  /**
-   * @return the unit
-   */
-  public int getUnit() {
-    return unit;
-  }
-
-  /**
-   * @param unit the unit to set
-   */
-  public void setUnit(int unit) {
-    this.unit = unit;
-  }
-
-  /**
-   * @return the price
-   */
-  public double getPrice() {
-    return price;
-  }
-
-  /**
-   * @param price the price to set
-   */
-  public void setPrice(double price) {
-    this.price = price;
-  }
 }
