@@ -1,16 +1,7 @@
 package com.bubuwork.jf.om.entity;
 
+import jakarta.persistence.*;
 import lombok.Data;
-
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
 
 @Entity(name="SALES_ORDER_ITEM")
 @Table
@@ -18,7 +9,7 @@ import javax.persistence.Table;
 public class OrderItem {
   
   @Id
-  @GeneratedValue(strategy=GenerationType.AUTO)
+  @GeneratedValue(strategy= GenerationType.AUTO)
   @Column(name="ID")
   private Long id;
   
